feat(path): Isolate Scoop apps' PATH

+++ lib/install.ps1
@@ -909,6 +909,12 @@ function env_add_path($manifest, $dir, $global, $arch) {
     $env_add_path = arch_specific 'env_add_path' $manifest $arch
     $dir = $dir.TrimEnd('\')
     if ($env_add_path) {
+        if (get_config USE_ISOLATED_PATH $false) {
+            Add-Path -Path '%SCOOP_PATH%' -Global:$global
+            $target_path = 'SCOOP_PATH'
+        } else {
+            $target_path = 'PATH'
+        }
         # GH-3785: Add path in ascending order.
         [Array]::Reverse($env_add_path)
         $env_add_path | Where-Object { $_ } | ForEach-Object {
@@ -920,7 +926,7 @@ function env_add_path($manifest, $dir, $global, $arch) {
             if (!(is_in_dir $dir $path_dir)) {
                 abort "Error in manifest: env_add_path '$_' is outside the app directory."
             }
-            Add-Path -Path $path_dir -Global:$global -Force
+            Add-Path -Path $path_dir -Scope $target_path -Global:$global -Force
         }
     }
 }
@@ -929,13 +935,21 @@ function env_rm_path($manifest, $dir, $global, $arch) {
     $env_add_path = arch_specific 'env_add_path' $manifest $arch
     $dir = $dir.TrimEnd('\')
     if ($env_add_path) {
+        if (get_config USE_ISOLATED_PATH $false) {
+            $target_path = 'SCOOP_PATH'
+        } else {
+            $target_path = 'PATH'
+        }
         $env_add_path | Where-Object { $_ } | ForEach-Object {
             if ($_ -eq '.') {
                 $path_dir = $dir
             } else {
                 $path_dir = Join-Path $dir $_
             }
-            Remove-Path -Path $path_dir -Global:$global
+            Remove-Path -Path $path_dir -Scope $target_path -Global:$global
+            if ($target_path -eq 'SCOOP_PATH') {
+                Remove-Path -Path $path_dir -Global:$global
+            }
         }
     }
 }
